extract container config preflight
